# Avoid allocations in more cases