@davidfowl 2019-04-26
What allocations are left? Just the selector context?