@rynowak 2019-04-26
> What allocations are left? Just the selector context?

Selector context - AND the `CandiateState[]` that we allocate.